telegram-selfie-bot/requirements.txt
python-telegram-bot==20.7
rembg==2.0.67
Pillow==10.4.0
gunicorn==21.2.0